NP-44903 Move license types
