+++ src/pages/public_registration/public_files/FileRow.tsx
@@ -18,10 +18,11 @@ import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import { downloadPrivateFile, downloadPublicFile } from '../../../api/fileApi';
 import { setNotification } from '../../../redux/notificationSlice';
 import { RootState } from '../../../redux/store';
-import { AssociatedFile, licenses } from '../../../types/associatedArtifact.types';
+import { AssociatedFile } from '../../../types/associatedArtifact.types';
 import { dataTestId } from '../../../utils/dataTestIds';
 import { PreviewFile } from './preview_file/PreviewFile';
 import { equalUris } from '../../../utils/general-helpers';
+import { licenses } from '../../../types/license.types';
 
 interface FileRowProps {
   file: AssociatedFile;
