+++ src/translations/enTranslations.json
@@ -475,7 +475,7 @@
       "CC BY-ND": "CC BY-ND ({{version}})",
       "CC BY-ND-NC": "CC BY-ND-NC ({{version}})",
       "CC BY-SA": "CC BY-SA ({{version}})",
-      "CC0": "CC0",
+      "CC0": "CC0 ({{version}})",
       "RightsReserved": "Publisher Terms"
     },
     "links": {
