+++ src/translations/nbTranslations.json
@@ -474,7 +474,7 @@
       "CC BY-ND": "CC BY-ND ({{version}})",
       "CC BY-ND-NC": "CC BY-ND-NC ({{version}})",
       "CC BY-SA": "CC BY-SA ({{version}})",
-      "CC0": "CC0",
+      "CC0": "CC0 ({{version}})",
       "RightsReserved": "Utgivers betingelser"
     },
     "links": {
