+++ src/types/associatedArtifact.types.ts
@@ -1,253 +1,4 @@
 import { Uppy as UppyType } from '@uppy/core';
-import * as LicenseImages from '../resources/images/licenses';
-import i18n from '../translations/i18n';
-
-enum LicenseUri {
-  CC_BY_4 = 'https://creativecommons.org/licenses/by/4.0',
-  CC_BY_3 = 'https://creativecommons.org/licenses/by/3.0',
-  CC_BY_2 = 'https://creativecommons.org/licenses/by/2.0',
-  CC_BY_1 = 'https://creativecommons.org/licenses/by/1.0',
-  CC_BY_SA_4 = 'https://creativecommons.org/licenses/by-sa/4.0',
-  CC_BY_SA_3 = 'https://creativecommons.org/licenses/by-sa/3.0',
-  CC_BY_SA_2 = 'https://creativecommons.org/licenses/by-sa/2.0',
-  CC_BY_SA_1 = 'https://creativecommons.org/licenses/by-sa/1.0',
-  CC_BY_ND_4 = 'https://creativecommons.org/licenses/by-nd/4.0',
-  CC_BY_ND_3 = 'https://creativecommons.org/licenses/by-nd/3.0',
-  CC_BY_ND_2 = 'https://creativecommons.org/licenses/by-nd/2.0',
-  CC_BY_ND_1 = 'https://creativecommons.org/licenses/by-nd/1.0',
-  CC_BY_NC_4 = 'https://creativecommons.org/licenses/by-nc/4.0',
-  CC_BY_NC_3 = 'https://creativecommons.org/licenses/by-nc/3.0',
-  CC_BY_NC_2 = 'https://creativecommons.org/licenses/by-nc/2.0',
-  CC_BY_NC_1 = 'https://creativecommons.org/licenses/by-nc/1.0',
-  CC_BY_NC_SA_4 = 'https://creativecommons.org/licenses/by-nc-sa/4.0',
-  CC_BY_NC_SA_3 = 'https://creativecommons.org/licenses/by-nc-sa/3.0',
-  CC_BY_NC_SA_2 = 'https://creativecommons.org/licenses/by-nc-sa/2.0',
-  CC_BY_NC_SA_1 = 'https://creativecommons.org/licenses/by-nc-sa/1.0',
-  CC_BY_NC_ND_4 = 'https://creativecommons.org/licenses/by-nc-nd/4.0',
-  CC_BY_NC_ND_3 = 'https://creativecommons.org/licenses/by-nc-nd/3.0',
-  CC_BY_NC_ND_2 = 'https://creativecommons.org/licenses/by-nc-nd/2.0',
-  CC_BY_ND_NC_1 = 'https://creativecommons.org/licenses/by-nd-nc/1.0',
-  CC0 = 'https://creativecommons.org/publicdomain/zero/1.0',
-  RightsReserved = 'https://rightsstatements.org/vocab/InC/1.0/',
-}
-
-interface LicenseInfo {
-  id: LicenseUri;
-  name: string;
-  description: string;
-  logo: string;
-  link: string;
-  version?: 1 | 2 | 3 | 4;
-}
-
-export const licenses: LicenseInfo[] = [
-  {
-    id: LicenseUri.CC_BY_4,
-    name: i18n.t('licenses.labels.CC BY', { version: '4.0' }),
-    description: i18n.t('licenses.description.cc_by'),
-    link: i18n.t('licenses.links.cc_by', { version: '4.0' }),
-    logo: LicenseImages.ccByLogo,
-    version: 4,
-  },
-  {
-    id: LicenseUri.CC_BY_3,
-    name: i18n.t('licenses.labels.CC BY', { version: '3.0' }),
-    description: i18n.t('licenses.description.cc_by'),
-    link: i18n.t('licenses.links.cc_by', { version: '3.0' }),
-    logo: LicenseImages.ccByLogo,
-    version: 3,
-  },
-  {
-    id: LicenseUri.CC_BY_2,
-    name: i18n.t('licenses.labels.CC BY', { version: '2.0' }),
-    description: i18n.t('licenses.description.cc_by'),
-    link: i18n.t('licenses.links.cc_by', { version: '2.0' }),
-    logo: LicenseImages.ccByLogo,
-    version: 2,
-  },
-  {
-    id: LicenseUri.CC_BY_1,
-    name: i18n.t('licenses.labels.CC BY', { version: '1.0' }),
-    description: i18n.t('licenses.description.cc_by'),
-    link: i18n.t('licenses.links.cc_by', { version: '1.0' }),
-    logo: LicenseImages.ccByLogo,
-    version: 1,
-  },
-  {
-    id: LicenseUri.CC_BY_SA_4,
-    name: i18n.t('licenses.labels.CC BY-SA', { version: '4.0' }),
-    description: i18n.t('licenses.description.cc_by_sa'),
-    link: i18n.t('licenses.links.cc_by_sa', { version: '4.0' }),
-    logo: LicenseImages.ccBySaLogo,
-    version: 4,
-  },
-  {
-    id: LicenseUri.CC_BY_SA_3,
-    name: i18n.t('licenses.labels.CC BY-SA', { version: '3.0' }),
-    description: i18n.t('licenses.description.cc_by_sa'),
-    link: i18n.t('licenses.links.cc_by_sa', { version: '3.0' }),
-    logo: LicenseImages.ccBySaLogo,
-    version: 3,
-  },
-  {
-    id: LicenseUri.CC_BY_SA_2,
-    name: i18n.t('licenses.labels.CC BY-SA', { version: '2.0' }),
-    description: i18n.t('licenses.description.cc_by_sa'),
-    link: i18n.t('licenses.links.cc_by_sa', { version: '2.0' }),
-    logo: LicenseImages.ccBySaLogo,
-    version: 2,
-  },
-  {
-    id: LicenseUri.CC_BY_SA_1,
-    name: i18n.t('licenses.labels.CC BY-SA', { version: '1.0' }),
-    description: i18n.t('licenses.description.cc_by_sa'),
-    link: i18n.t('licenses.links.cc_by_sa', { version: '1.0' }),
-    logo: LicenseImages.ccBySaLogo,
-    version: 1,
-  },
-  {
-    id: LicenseUri.CC_BY_ND_4,
-    name: i18n.t('licenses.labels.CC BY-ND', { version: '4.0' }),
-    description: i18n.t('licenses.description.cc_by_nd'),
-    link: i18n.t('licenses.links.cc_by_nd', { version: '4.0' }),
-    logo: LicenseImages.ccByNdLogo,
-    version: 4,
-  },
-  {
-    id: LicenseUri.CC_BY_ND_3,
-    name: i18n.t('licenses.labels.CC BY-ND', { version: '3.0' }),
-    description: i18n.t('licenses.description.cc_by_nd'),
-    link: i18n.t('licenses.links.cc_by_nd', { version: '3.0' }),
-    logo: LicenseImages.ccByNdLogo,
-    version: 3,
-  },
-  {
-    id: LicenseUri.CC_BY_ND_2,
-    name: i18n.t('licenses.labels.CC BY-ND', { version: '2.0' }),
-    description: i18n.t('licenses.description.cc_by_nd'),
-    link: i18n.t('licenses.links.cc_by_nd', { version: '2.0' }),
-    logo: LicenseImages.ccByNdLogo,
-    version: 2,
-  },
-  {
-    id: LicenseUri.CC_BY_ND_1,
-    name: i18n.t('licenses.labels.CC BY-ND', { version: '1.0' }),
-    description: i18n.t('licenses.description.cc_by_nd'),
-    link: i18n.t('licenses.links.cc_by_nd', { version: '1.0' }),
-    logo: LicenseImages.ccByNdLogo,
-    version: 1,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_4,
-    name: i18n.t('licenses.labels.CC BY-NC', { version: '4.0' }),
-    description: i18n.t('licenses.description.cc_by_nc'),
-    link: i18n.t('licenses.links.cc_by_nc', { version: '4.0' }),
-    logo: LicenseImages.ccByNcLogo,
-    version: 4,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_3,
-    name: i18n.t('licenses.labels.CC BY-NC', { version: '3.0' }),
-    description: i18n.t('licenses.description.cc_by_nc'),
-    link: i18n.t('licenses.links.cc_by_nc', { version: '3.0' }),
-    logo: LicenseImages.ccByNcLogo,
-    version: 3,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_2,
-    name: i18n.t('licenses.labels.CC BY-NC', { version: '2.0' }),
-    description: i18n.t('licenses.description.cc_by_nc'),
-    link: i18n.t('licenses.links.cc_by_nc', { version: '2.0' }),
-    logo: LicenseImages.ccByNcLogo,
-    version: 2,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_1,
-    name: i18n.t('licenses.labels.CC BY-NC', { version: '1.0' }),
-    description: i18n.t('licenses.description.cc_by_nc'),
-    link: i18n.t('licenses.links.cc_by_nc', { version: '1.0' }),
-    logo: LicenseImages.ccByNcLogo,
-    version: 1,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_SA_4,
-    name: i18n.t('licenses.labels.CC BY-NC-SA', { version: '4.0' }),
-    description: i18n.t('licenses.description.cc_by_nc_sa'),
-    link: i18n.t('licenses.links.cc_by_nc_sa', { version: '4.0' }),
-    logo: LicenseImages.ccByNcSaLogo,
-    version: 4,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_SA_3,
-    name: i18n.t('licenses.labels.CC BY-NC-SA', { version: '3.0' }),
-    description: i18n.t('licenses.description.cc_by_nc_sa'),
-    link: i18n.t('licenses.links.cc_by_nc_sa', { version: '3.0' }),
-    logo: LicenseImages.ccByNcSaLogo,
-    version: 3,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_SA_2,
-    name: i18n.t('licenses.labels.CC BY-NC-SA', { version: '2.0' }),
-    description: i18n.t('licenses.description.cc_by_nc_sa'),
-    link: i18n.t('licenses.links.cc_by_nc_sa', { version: '2.0' }),
-    logo: LicenseImages.ccByNcSaLogo,
-    version: 2,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_SA_1,
-    name: i18n.t('licenses.labels.CC BY-NC-SA', { version: '1.0' }),
-    description: i18n.t('licenses.description.cc_by_nc_sa'),
-    link: i18n.t('licenses.links.cc_by_nc_sa', { version: '1.0' }),
-    logo: LicenseImages.ccByNcSaLogo,
-    version: 1,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_ND_4,
-    name: i18n.t('licenses.labels.CC BY-NC-ND', { version: '4.0' }),
-    description: i18n.t('licenses.description.cc_by_nc_nd'),
-    link: i18n.t('licenses.links.cc_by_nc_nd', { version: '4.0' }),
-    logo: LicenseImages.ccByNcNdLogo,
-    version: 4,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_ND_3,
-    name: i18n.t('licenses.labels.CC BY-NC-ND', { version: '3.0' }),
-    description: i18n.t('licenses.description.cc_by_nc_nd'),
-    link: i18n.t('licenses.links.cc_by_nc_nd', { version: '3.0' }),
-    logo: LicenseImages.ccByNcNdLogo,
-    version: 3,
-  },
-  {
-    id: LicenseUri.CC_BY_NC_ND_2,
-    name: i18n.t('licenses.labels.CC BY-NC-ND', { version: '2.0' }),
-    description: i18n.t('licenses.description.cc_by_nc_nd'),
-    link: i18n.t('licenses.links.cc_by_nc_nd', { version: '2.0' }),
-    logo: LicenseImages.ccByNcNdLogo,
-    version: 2,
-  },
-  {
-    id: LicenseUri.CC_BY_ND_NC_1,
-    name: i18n.t('licenses.labels.CC BY-ND-NC', { version: '1.0' }),
-    description: i18n.t('licenses.description.cc_by_nc_nd'),
-    link: i18n.t('licenses.links.cc_by_nd_nc', { version: '1.0' }),
-    logo: LicenseImages.ccByNcNdLogo,
-    version: 1,
-  },
-  {
-    id: LicenseUri.CC0,
-    name: i18n.t('licenses.labels.CC0'),
-    description: i18n.t('licenses.description.cc0'),
-    link: i18n.t('licenses.links.cc0'),
-    logo: LicenseImages.cc0Logo,
-  },
-  {
-    id: LicenseUri.RightsReserved,
-    name: i18n.t('licenses.labels.RightsReserved'),
-    description: i18n.t('licenses.description.rights_reserved'),
-    link: i18n.t('licenses.links.rights_reserved'),
-    logo: LicenseImages.rightsReservedLogo,
-  },
-];
 
 export type AssociatedFileType = 'PublishedFile' | 'UnpublishedFile' | 'UnpublishableFile';
